# api/openapi.yaml
openapi: 3.1.0
info:
  title: CineForge Control API
  version: 0.1.0
  description: Account, workspace, project, library, media-ingestion, and audit foundation.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/sessions:
    get:
      tags: [Auth]
      operationId: listSessions
      responses:
        '200':
          description: Active account sessions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionList' }
        default: { $ref: '#/components/responses/Problem' }
  /auth/sessions/{sessionId}:
    delete:
      tags: [Auth]
      operationId: revokeSession
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Revoked }
        default: { $ref: '#/components/responses/Problem' }
...
  /auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: All sessions revoked }
        default: { $ref: '#/components/responses/Problem' }
...
  /me/deletion:
    post:
      tags: [Account]
      operationId: scheduleAccountDeletion
      description: Requires recent authentication and an idempotency key.
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202':
          description: Deletion scheduled.
          content: { application/json: { schema: { $ref: '#/components/schemas/AccountDeletionRequest' } } }
        default: { $ref: '#/components/responses/Problem' }
  /me/deletion/cancel:
    post:
      tags: [Account]
      operationId: cancelAccountDeletion
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Deletion cancelled }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [active, trashed] } }
      responses:
        '200':
          description: Workspaces visible to the caller.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Workspace' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 2, maxLength: 120 }
      responses:
        '201':
          description: Team workspace created.
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/bootstrap:
    post:
      tags: [Workspaces]
      operationId: bootstrapPersonalWorkspace
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '200':
          description: Existing or newly created personal workspace.
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}:
    patch:
      tags: [Workspaces]
      operationId: updateWorkspace
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/IfMatch' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string, minLength: 2, maxLength: 120 } }
      responses:
        '200':
          description: Workspace updated.
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Workspaces]
      operationId: trashWorkspace
      description: Requires recent authentication and an idempotency key.
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '202': { description: Workspace moved to trash lifecycle }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/restore:
    post:
      tags: [Workspaces]
      operationId: restoreWorkspace
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Workspace restored }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/members:
    get:
      tags: [Workspaces]
      operationId: listMembers
      parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
      responses:
        '200':
          description: Workspace memberships.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Membership' }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/members/{memberId}:
    patch:
      tags: [Workspaces]
      operationId: updateMember
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/MemberId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MemberUpdate' } } }
      responses:
        '200':
          description: Updated membership.
          content: { application/json: { schema: { $ref: '#/components/schemas/Membership' } } }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/invitations:
    get:
      tags: [Workspaces]
      operationId: listInvitations
      parameters: [{ $ref: '#/components/parameters/WorkspaceId' }]
      responses:
        '200':
          description: Pending invitations.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Invitation' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Workspaces]
      operationId: createInvitation
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/InvitationCreate' } } }
      responses:
        '201':
          description: Invitation created and mailed.
          content: { application/json: { schema: { $ref: '#/components/schemas/Invitation' } } }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/invitations/{invitationId}:
    delete:
      tags: [Workspaces]
      operationId: revokeInvitation
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Invitation revoked }
        default: { $ref: '#/components/responses/Problem' }
  /invitations/accept:
    post:
      tags: [Workspaces]
      operationId: acceptInvitation
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Invitation accepted.
          content: { application/json: { schema: { $ref: '#/components/schemas/Workspace' } } }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/ownership:
    post:
      tags: [Workspaces]
      operationId: transferOwnership
      description: Requires recent authentication.
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId]
              properties:
                memberId: { type: string, format: uuid }
      responses:
        '204': { description: Ownership transferred }
        default: { $ref: '#/components/responses/Problem' }
...
  /workspaces/{workspaceId}/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { name: status, in: query, schema: { type: string, enum: [active, archived, trashed] } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: Projects.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Projects]
      operationId: createProject
      parameters:
        - { $ref: '#/components/parameters/WorkspaceId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectCreateInput' } } }
      responses:
        '201':
          description: Project created.
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        default: { $ref: '#/components/responses/Problem' }
  /projects/{projectId}:
    get:
      tags: [Projects]
      operationId: getProject
      parameters: [{ $ref: '#/components/parameters/ProjectId' }]
      responses:
        '200':
          description: Project.
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        default: { $ref: '#/components/responses/Problem' }
    put:
      tags: [Projects]
      operationId: updateProject
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/IfMatch' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectSettings' } } }
      responses:
        '200':
          description: Project updated.
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Projects]
      operationId: trashProject
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Project moved to trash.
          content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: cf_session }
  parameters:
    WorkspaceId: { name: workspaceId, in: path, required: true, schema: { type: string, format: uuid } }
...
    SessionId: { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string } }
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 128 } }
  responses:
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    SessionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AccountSession' }
    AccountDeletionRequest:
      type: object
      required: [scheduledAt, purgeAt]
      properties:
        scheduledAt: { type: string, format: date-time }
        purgeAt: { type: string, format: date-time }
    Workspace:
      type: object
      required: [id, name, slug, kind, ownerUserId, version]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        kind: { type: string, enum: [personal, team] }
        ownerUserId: { type: string, format: uuid }
        status: { type: string, enum: [active, trashed], default: active }
        version: { type: integer, format: int64 }
        deletedAt: { type: [string, 'null'], format: date-time }
        purgeAt: { type: [string, 'null'], format: date-time }
        currentMembership: { $ref: '#/components/schemas/Membership' }
...